python/requirements.txt
libusb1>=1.4.1
